Refactor BrokerGrpcFacade to return ListenableFutures
